Split natural-language orchestration into focused services
