+++ src/fred_query/services/__init__.py
@@ -1,11 +1,14 @@
 from fred_query.services.answer_service import AnswerService
 from fred_query.services.chart_service import ChartService
+from fred_query.services.clarification_resolver import ClarificationResolver
 from fred_query.services.comparison_service import StateGDPComparisonService
 from fred_query.services.cross_section_service import CrossSectionService
 from fred_query.services.fred_client import FREDAPIError, FREDClient
+from fred_query.services.follow_up_intent_merger import FollowUpIntentMerger
 from fred_query.services.intent_service import IntentService
 from fred_query.services.natural_language_query_service import NaturalLanguageQueryService
 from fred_query.services.openai_parser_service import OpenAIIntentParser
+from fred_query.services.query_router import QueryRouter
 from fred_query.services.query_session_service import QuerySession, QuerySessionService
 from fred_query.services.relationship_service import RelationshipAnalysisService
 from fred_query.services.resolver_service import ResolverService
@@ -15,12 +18,15 @@ from fred_query.services.transform_service import TransformService
 __all__ = [
     "AnswerService",
     "ChartService",
+    "ClarificationResolver",
     "CrossSectionService",
     "FREDAPIError",
     "FREDClient",
+    "FollowUpIntentMerger",
     "IntentService",
     "NaturalLanguageQueryService",
     "OpenAIIntentParser",
+    "QueryRouter",
     "QuerySession",
     "QuerySessionService",
     "RelationshipAnalysisService",
